backend/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sse-starlette>=2.2.0
claude-agent-sdk>=0.1.45
yfinance>=1.2.0
requests>=2.32.0
python-dotenv>=1.2.0
